+++ app/ui/latest/src/main/java/com/mudita/mail/repository/providers/model/ProviderTile.kt
@@ -1,7 +1,9 @@
 package com.mudita.mail.repository.providers.model
 
+import com.mudita.mail.ui.util.stringKey.StringKey
+
 data class ProviderTile(
-    val name: String,
-    val description: String,
+    val nameKey: StringKey,
+    val descriptionKey: StringKey,
     val type: ProviderType
 )
